python/infer/requirements.txt
absl-py==2.2.2
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
astunparse==1.6.3
attrs==25.3.0
bleach==6.2.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
deprecation==2.1.0
dotenv==0.9.9
flatbuffers==25.2.10
frozenlist==1.5.0
gast==0.6.0
google-auth==2.38.0
google-auth-oauthlib==1.2.1
google-genai==1.10.0
google-pasta==0.2.0
gotrue==2.12.0
grpcio==1.71.0
h11==0.14.0
h2==4.2.0
h5py==3.13.0
hpack==4.1.0
httpcore==1.0.7
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
iniconfig==2.1.0
joblib==1.4.2
kaggle==1.7.4.2
kagglehub==0.3.11
keras==2.15.0
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml-dtypes==0.3.2
multidict==6.2.0
namex==0.0.8
numpy==1.26.4
oauthlib==3.2.2
opt_einsum==3.4.0
optree==0.14.1
packaging==24.2
pluggy==1.5.0
postgrest==1.0.1
propcache==0.3.1
protobuf==4.25.6
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.1
pydantic_core==2.33.0
Pygments==2.19.1
PyJWT==2.10.1
pytest==8.3.5
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-slugify==8.0.4
PyYAML==6.0.2
realtime==2.4.2
requests==2.32.3
requests-oauthlib==2.0.0
rich==14.0.0
rsa==4.9
scikit-learn==1.6.1
scipy==1.15.2
six==1.17.0
sniffio==1.3.1
storage3==0.11.3
StrEnum==0.4.15
supabase==2.15.0
supafunc==0.9.4
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow==2.15.1
tensorflow-estimator==2.15.0
tensorflow-hub==0.16.1
tensorflow-io-gcs-filesystem==0.37.1
tensorflow-text==2.15.0
termcolor==3.0.1
text-unidecode==1.3
tf_keras==2.15.1
threadpoolctl==3.6.0
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.0
urllib3==2.3.0
webencodings==0.5.1
websockets==14.2
Werkzeug==3.1.3
wrapt==1.14.1
yarl==1.18.3